Add step_time field in StateEstimation.msg

+++ ros2/src/march_state_estimator/include/march_state_estimator/state_estimator_node.hpp
@@ -22,7 +22,7 @@ private:
     void imuCallback(const sensor_msgs::msg::Imu::SharedPtr msg);
     void publishStateEstimation();
 
-    int64_t m_dt;
+    double m_dt;
     sensor_msgs::msg::JointState::SharedPtr m_joint_state;
     sensor_msgs::msg::Imu::SharedPtr m_imu;
